Advance deprecation level of FP to lesser than Int types to ERROR #KT-30360
(cherry picked from commit 968099fbec46472b0b18809698384ba1f4b16c33)

+++ j2k/testData/fileOrElement/methodCallExpression/specialBuiltinMembers.kt
@@ -22,7 +22,7 @@ internal class A {
         val i = 10.1.toInt()
         val f = 10.1.toFloat()
         val l = 10.1.toLong()
-        val s = 10.1.toShort()
+        val s = 10.1.toInt().toShort()
 
         try {
             val removed = list.removeAt(10)
